@@ -13,44 +13,37 @@
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write .",
 		"test:integration": "playwright test",
-		"test:unit": "vitest",
-		"postinstall": "prisma generate"
+		"test:unit": "vitest"
 	},
 	"devDependencies": {
-		"@iconify/svelte": "^3.1.6",
+		"@iconify/svelte": "^4.1.0",
 		"@playwright/test": "^1.28.1",
-		"@sveltejs/adapter-vercel": "^5.1.1",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/eslint": "8.56.0",
+		"@sveltejs/adapter-vercel": "^5.5.2",
+		"@sveltejs/kit": "^2.15.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@types/eslint": "9.6.1",
 		"@types/nodemailer": "^6.4.14",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"autoprefixer": "^10.4.17",
-		"eslint": "^8.56.0",
+		"@typescript-eslint/eslint-plugin": "^8.18.2",
+		"@typescript-eslint/parser": "^8.18.2",
+		"autoprefixer": "^10.4.20",
+		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"postcss": "^8.4.35",
+		"postcss": "^8.4.49",
 		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.5.11",
-		"prisma": "^5.13.0",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.4.1",
-		"tslib": "^2.4.1",
+		"prettier-plugin-tailwindcss": "^0.6.9",
+		"prisma": "^6.1.0",
+		"svelte": "^5.16.0",
+		"tailwindcss": "^3.4.17",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
+		"vite": "^6.0.6",
 		"vitest": "^1.2.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"@prisma/client": "^5.13.0",
-		"@prisma/extension-accelerate": "^1.0.0",
+		"@prisma/client": "^6.1.0",
 		"@stripe/stripe-js": "^3.0.10",
-		"@supabase/ssr": "^0.1.0",
-		"@supabase/supabase-js": "^2.39.8",
 		"base64-arraybuffer": "^1.0.2",
+		"better-auth": "^1.1.4",
 		"stripe": "^14.21.0",
 		"svelte-dnd-action": "^0.9.42",
 		"tailwind-merge": "^2.2.1"
