@@ -1,7 +1,17 @@
-import type { LayoutServerLoad } from './$types';
+// import type { LayoutServerLoad } from "./$types";
 
-export const load: LayoutServerLoad = async ({ locals: { session } }) => {
-	return {
-		session
-	};
-};
+// export const load: LayoutServerLoad = async ({ locals }) => {
+//   const user = locals.session?.passport?.user;
+
+//   return {
+//     session: {
+//       user: user
+//         ? {
+//             id: user.id,
+//             email: user.email,
+//             role: user.role,
+//           }
+//         : null,
+//     },
+//   };
+// };
