@@ -1,11 +1,12 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-	content: ['./src/**/*.{html,js,svelte,ts}'],
-	theme: {
-		extend: {
-			fontFamily: {
-				norwester: ['Norwester']
-			}
-		}
-	}
+  content: ['./src/**/*.{html,js,svelte,ts}'],
+  theme: {
+    extend: {
+      fontFamily: {
+        norwester: ['Norwester']
+      }
+    }
+  },
+  plugins: []
 };
